Added some major styling and conditional rendering.
Issues are whether or not to have separate article divs for rendering.

Styles.js is now getting to be bogged down. Refactoring will be needed
in order to accommodate growing file load.

Portfolio and Contact pages must be made in XD. Further image assets
to be added.

Problems with CSS inheritance could be an issue RE: color scheming.
Portfolio and About pages will probably be the same, unless a pop-up
full screen modal will do for the contact page.